Added function implementation
Added implementation of countNumbers(...) function. This implementation starts at the first element of the sorted vector, than traverses the vector until either we reach an element whose value is greater than the given parameter or we have reached the end of the vector.

Although this implementation provides correct results, it is taking too long to execute. I need to find a more efficient alternative to the while(...) loop.