Added TrainNode class
I've created a TrainNode class that describes each train in the TrainComposition. Since the functionality of the TrainComposition is basically that of a doubly-linked list, this TrainNode class is virtually identical to that of a node for any dll. 

Normally I would create a separate file for this TrainNode class to keep my code organized, but for ease of copying/pasting to testdome.com I have decided to add it in the same file as the TrainComposition class.