Got rid of "using namespace std"
For the sake of consistency among files I removed the "using namespace std" statement from the beginning of the file. This also means that I needed to change instances of "string" to "std::string". 

This obviously does not affect the functionality of the code, but since I did not include the "using namespace std" statement in any of the other files in this repository I figured I should make this one match the rest.